Renamed package to reflect endpoint root
